requirements.txt
aws-cdk-lib==2.100.0
aws-cdk.aws-lambda-python-alpha==2.100.0a0
constructs>=10.0.0,<11.0.0
boto3>=1.34.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0

# Code quality
flake8>=6.1.0
black>=23.12.0

# Additional dependencies for Python 3.12 compatibility
typing-extensions>=4.8.0
